Improved autocomplete styling
