+++ client/src/components/FormFields/Autocomplete.scss
@@ -1,42 +1,61 @@
 .autocomplete {
   width: 100%;
-  height: 25px;
+  font-size: 14px;
+  color: black;
   border-radius: 4px;
+  background-color: white;
+  position: relative;
 
-  .autoContainer {
+  .searchContainer {
     width: 100%;
-    font-size: 14px;
-    color: black;
-    background-color: white;
-    border: 1px solid black;
-    border-radius: 4px;
+    padding: 1px;
 
-    .tag {
-      width: fit-content;
-      display: inline-flex;
-      align-items: center;
-      gap: 5px;
-      background-color: var(--color-base);
-      color: white;
-      padding: 4px;
-      border-radius: 4px;
-      margin-right: 3px;
-
-      .value {
-        font-size: 12px;
-      }
-
-      .x:hover {
-        cursor: pointer;
-      }
-    }
-
-    .search {
+    .searchbar {
       margin: 0;
       padding: 4px 8px;
       border: none;
       background-color: transparent;
       font-size: inherit;
     }
+  }
+
+  &.multiple {
+    .searchContainer {
+      display: flex;
+      align-items: center;
+      flex-wrap: wrap;
+      gap: 2px;
+
+      .tag {
+        width: fit-content;
+        display: inline-flex;
+        align-items: center;
+        gap: 5px;
+        background-color: var(--color-base);
+        color: white;
+        padding: 4px;
+        border-radius: 4px;
+
+        .value {
+          font-size: 12px;
+        }
+
+        .x:hover {
+          cursor: pointer;
+        }
+      }
+
+      .searchbar {
+        min-width: 100px;
+        flex: 1;
+      }
+    }
+  }
+
+  .dropdownMenu {
+    position: absolute;
+    top: calc(100% + 1px);
+    z-index: 100;
+    margin: 1px 0;
   }
 }
